LLM/ChatPDF/requirements.txt
langchain
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
PyPDF2